+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Braintree Android Drop-In Release Notes
 
+## unreleased
+
+* Add California Privacy Laws notice of collection to credit card form
+
 ## 6.10.0
 
 * Bump braintree_android module dependency versions to `4.33.0`